+++ ValidCode/Abstract.cs
@@ -1,0 +1,9 @@
+﻿namespace ValidCode
+{
+    using System;
+
+    public abstract class Abstract : IDisposable
+    {
+        public abstract void Dispose();
+    }
+}