+++ ValidCode/Locals.cs
@@ -1,4 +1,4 @@
-namespace ValidCode
+﻿namespace ValidCode
 {
     using System.IO;