+++ RELEASE_NOTES.md
@@ -1,3 +1,4 @@
+* BUGFIX: IDISP025 when abstract dispose method.
 * BUGFIX: IDISP006 when explicit implementation.
 
 #### 3.3.0